Added empty state card
