+++ plugins/humanitec/src/annotations.ts
@@ -1,0 +1,4 @@
+export const HUMANITEC_ORG_ID_ANNOTATION = 'humanitec.com/orgId';
+export const HUMANITEC_APP_ID_ANNOTATION = 'humanitec.com/appId';
+
+export const HUMANITEC_MISSING_ANNOTATION_ERROR = 'Entity is missing Humanitec annotations.';
