+++ plugins/humanitec/src/components/HumanitecCardComponent.tsx
@@ -1,13 +1,15 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { useEntity } from '@backstage/plugin-catalog-react';
-import { HumanitecCard } from './HumanitecCard';
 import { useAppInfo } from '../hooks/useAppInfo';
 import { HumanitecAnnotationedEntity } from '../types';
 import { useHumanitecParams } from '../hooks/useHumanitecParams';
 import { useStyles } from '../hooks/useStyles';
 import CardHeader from '@material-ui/core/CardHeader';
-import { Button } from '@material-ui/core';
+import { Button, Card, CardContent, Divider } from '@material-ui/core';
 import { HumanitecLogoIcon } from './HumanitecLogoIcon';
+import { HUMANITEC_APP_ID_ANNOTATION, HUMANITEC_MISSING_ANNOTATION_ERROR, HUMANITEC_ORG_ID_ANNOTATION } from '../annotations';
+import { HumanitecCardContent } from './HumanitecCardContent';
+import { HumanitecAnnotationsEmptyState } from './HumanitecAnnotationsEmptyState';
 
 interface HumanitecCardComponentProps {
   variant: 'gridItem'
@@ -16,8 +18,8 @@ interface HumanitecCardComponentProps {
 export function HumanitecCardComponent({ variant }: HumanitecCardComponentProps) {
   const { entity } = useEntity<HumanitecAnnotationedEntity>();
 
-  const appId = entity.metadata.annotations['humanitec.com/appId'];
-  const orgId = entity.metadata.annotations['humanitec.com/orgId'];
+  const orgId = entity.metadata.annotations[HUMANITEC_ORG_ID_ANNOTATION];
+  const appId = entity.metadata.annotations[HUMANITEC_APP_ID_ANNOTATION];
   const appUrl = `https://app.humanitec.io/orgs/${orgId}/apps/${appId}`;
 
   const data = useAppInfo({
@@ -29,24 +31,47 @@ export function HumanitecCardComponent({ variant }: HumanitecCardComponentProps)
 
   const classes = useStyles();
 
-  return (<HumanitecCard
-    variant={variant}
-    header={
+  let cardContentClass = '';
+  if (variant === 'gridItem') {
+    cardContentClass = classes.gridItemCardContent;
+  }
+
+  let content: ReactNode = null;
+  if (Array.isArray(data)) {
+    content = (
+      <HumanitecCardContent
+        environments={data}
+        appUrl={appUrl}
+        selectedEnv={params?.envId}
+        selectedWorkload={params?.workloadId}
+        actions={actions}
+      />
+    )
+  } else if (data instanceof Error && data.message === HUMANITEC_MISSING_ANNOTATION_ERROR) {
+    content = (<HumanitecAnnotationsEmptyState />)
+  }
+
+  let action: ReactNode = null;
+  if (appId && orgId) {
+    action = (
+      <Button component="a" startIcon={<HumanitecLogoIcon />} href={appUrl}>
+        Humanitec App
+      </Button>
+    )
+  }
+
+  return (
+    <Card className={`${classes.cardClass} ${cardContentClass}`} >
       <CardHeader
-        action={
-          <Button component="a" startIcon={<HumanitecLogoIcon />} href={appUrl}>
-            Humanitec App
-          </Button>
-        }
+        action={action}
         className={classes.cardHeader}
         title="Deployments"
         subheader="Orchestrated by Humanitec"
       />
-    }
-    environments={data}
-    appUrl={appUrl}
-    selectedEnv={params?.envId}
-    selectedWorkload={params?.workloadId}
-    actions={actions}
-  />)
+      <Divider />
+      <CardContent className={classes.gridItemCardContent}>
+        {content}
+      </CardContent>
+    </Card>
+  )
 }
